chore: info panel
